refactor: proximity machine
