+++ ts/features/itwallet/common/utils/theme.ts
@@ -36,7 +36,9 @@ const themeKeys = [
   // Banners
   "banner-background",
   // Cards
-  "card-background"
+  "card-background",
+  // Proximity
+  "qrcode-color"
 ] as const;
 
 export type ItWalletTheme = {
@@ -46,13 +48,15 @@ export type ItWalletTheme = {
 const itWalletLightTheme: ItWalletTheme = {
   "header-background": "#F2F9FF",
   "banner-background": "#F2F9FF",
-  "card-background": "#F2F2F2"
+  "card-background": "#F2F2F2",
+  "qrcode-color": "#000000"
 };
 
 const itWalletDarkTheme: ItWalletTheme = {
   "header-background": "#0F2433",
   "banner-background": "#0F2433",
-  "card-background": "#1A1A1A"
+  "card-background": "#1A1A1A",
+  "qrcode-color": "#FFFFFF"
 };
 
 export const ItWalletThemes = {
