+++ ts/features/itwallet/presentation/proximity/machine/actors.ts
@@ -1,22 +1,10 @@
 import { ISO18013_5 } from "@pagopa/io-react-native-iso18013";
-import { Platform } from "react-native";
 import BluetoothStateManager from "react-native-bluetooth-state-manager";
-import {
-  checkMultiple,
-  Permission,
-  PERMISSIONS,
-  requestMultiple,
-  RESULTS
-} from "react-native-permissions";
 import { fromCallback, fromPromise } from "xstate";
 import { assert } from "../../../../../utils/assert";
 import { Env } from "../../../common/utils/environment";
 import { CredentialFormat } from "../../../common/utils/itwTypesUtils";
 import { CredentialsVault } from "../../../credentials/utils/vault";
-import {
-  trackItwProximityBluetoothBlock,
-  trackItwProximityBluetoothBlockAction
-} from "../analytics";
 import {
   generateAcceptedFields,
   getDocuments,
@@ -24,25 +12,11 @@ import {
   promiseWithTimeout
 } from "../utils/itwProximityPresentationUtils";
 import type { EventsPayload } from "../utils/itwProximityTypeUtils";
+import { checkBluetoothPermissions } from "../utils/permissions";
 import { Context } from "./context";
 import { ProximityEvents } from "./events";
 
-const PERMISSIONS_TO_CHECK: Array<Permission> =
-  Platform.OS === "android"
-    ? Platform.Version >= 31
-      ? [
-          PERMISSIONS.ANDROID.BLUETOOTH_SCAN,
-          PERMISSIONS.ANDROID.BLUETOOTH_CONNECT,
-          PERMISSIONS.ANDROID.BLUETOOTH_ADVERTISE
-        ] // Android 12 and above: Request new Bluetooth permissions along with location.
-      : [PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION] // Android 9 to Android 11: Only location permission is required for BLE.
-    : [PERMISSIONS.IOS.BLUETOOTH]; // iOS permissions required are Bluetooth and location.
-
 const SEND_RESPONSE_TIMEOUT_MS = 20000;
-
-export type CheckPermissionsInput = {
-  isSilent?: boolean;
-};
 
 export type SendErrorResponseActorOutput = Awaited<
   ReturnType<typeof ISO18013_5.sendErrorResponse>
@@ -62,44 +36,12 @@ export type SendDocumentsActorOutput = Awaited<
 >;
 
 export const createProximityActorsImplementation = (env: Env) => {
-  const checkPermissions = fromPromise<boolean, CheckPermissionsInput>(
-    async ({ input }) => {
-      const isSilent = input?.isSilent || false;
-
-      // Check current permission status
-      const statuses = await checkMultiple(PERMISSIONS_TO_CHECK);
-
-      // Filter out already granted permissions
-      const permissionsToRequest = PERMISSIONS_TO_CHECK.filter(
-        permission => statuses[permission] !== RESULTS.GRANTED
-      );
-
-      if (permissionsToRequest.length > 0) {
-        if (!isSilent) {
-          trackItwProximityBluetoothBlock();
-        }
-        // Request only the missing permissions
-        const requestResults = await requestMultiple(permissionsToRequest);
-
-        const allPermissionsGranted = permissionsToRequest.every(
-          permission => requestResults[permission] === RESULTS.GRANTED
-        );
-
-        if (!isSilent) {
-          const userAction = allPermissionsGranted ? "allow" : "not_allow";
-          trackItwProximityBluetoothBlockAction(userAction);
-        }
-
-        // Verify if all requested permissions are granted
-        return allPermissionsGranted;
-      }
-      return true;
-    }
+  const checkBluetoohPermissions = fromPromise<boolean>(
+    checkBluetoothPermissions
   );
 
-  const checkBluetoothIsActive = fromPromise<boolean, void>(async () => {
+  const checkBluetoothActivation = fromPromise<boolean>(async () => {
     const bluetoothState = await BluetoothStateManager.getState();
-
     return bluetoothState === "PoweredOn";
   });
 
@@ -250,8 +192,8 @@ export const createProximityActorsImplementation = (env: Env) => {
   );
 
   return {
-    checkPermissions,
-    checkBluetoothIsActive,
+    checkBluetoohPermissions,
+    checkBluetoothActivation,
     startEngagement,
     proximityCommunicationLogic,
     closeProximityFlow,
